+++ cmake/FindFreeImage.cmake
@@ -24,7 +24,7 @@ set(minor_version ${FreeImage_FIND_VERSION_MINOR})
 set(full_version ${major_version}.${minor_version})
 
 if (NOT WIN32)
-  include(IgnPkgConfig)
+  include(GzPkgConfig)
   gz_pkg_config_library_entry(FreeImage freeimage)
 
   # If we don't have PkgConfig, or if PkgConfig failed, then do a manual search
@@ -87,7 +87,7 @@ if (NOT WIN32)
 
   if(FreeImage_FOUND)
     # Create the imported target for FreeImage if we found it
-    include(IgnImportTarget)
+    include(GzImportTarget)
     gz_import_target(FreeImage)
   endif()
 
@@ -140,9 +140,9 @@ else()
   endif()
 
   if (FreeImage_FOUND)
-    include(IgnPkgConfig)
+    include(GzPkgConfig)
     gz_pkg_config_library_entry(FreeImage "FreeImage")
-    include(IgnImportTarget)
+    include(GzImportTarget)
     gz_import_target(FreeImage)
   endif()