+++ cmake/GzCodeCheck.cmake
@@ -1,7 +1,7 @@
 # Setup the codecheck target, which will run cppcheck and cppplint.
 # This function is private to gz-cmake.
 function(_gz_setup_target_for_codecheck)
-  include(IgnPython)
+  include(GzPython)
 
   find_program(CPPCHECK_PATH cppcheck)
   find_program(FIND_PATH find)