+++ cmake/GzManualSearch.cmake
@@ -107,7 +107,7 @@ macro(gz_manual_search package)
   endif()
 
   if(${package}_FOUND)
-    include(IgnImportTarget)
+    include(GzImportTarget)
     gz_import_target(${package} ${_gz_pkg_check_modules_interface_option}
       TARGET_NAME ${gz_pkg_check_modules_TARGET_NAME})
   endif()